requirements.txt
streamlit>=1.30.0
google-genai>=1.0.0
google-api-python-client>=2.100.0
youtube-transcript-api>=0.6.0
python-dotenv>=1.0.0
fpdf2>=2.7.0
requests>=2.31.0
